script to run a collection of queries on wikis with variable substitution
I needed this to audit the results of a bunch of queries on various
dbs and wikis, with different values for page ids, rev ids, page
ranges and titles depending on the wiki.

No one else will probably ever want it, and as you may notice I
wholesale copied the php file from the checktables directory
because lazy.

The sample-local-settings file is just the (redacted) copy of
what I have for one wiki on my local install.