truly horrid script to produce config stanzas for some show explains
Too many steps for me to want to do this by hand for a bunch of wikis
or even for one wiki multiple times. So here is this script.
It's only tested for dryrun, so likely all broken, but at least it's
committed :-P

Oh yeah and it needs this crap revsperpage utility, so I added that
too. Makefile? What's that?