move DbInfo out to a separate module
This lets us reuse get_dbhosts, get_dbcreds etc.
It also required moving build_command and related methods out to
a separate module, since it's used by DbInfo.
SSH handling had to be cleaned up; now the caller is responsible for
removing mwhost from config settings before they are passed in to
any classes that should run commands only locally and not remotely
on the mwhost.
dryrun and verbose got moved into the args dict to be passed around,
so that all the calling modules behave the same way.
sudouser is a config setting so it can be turned on or off depending
on one's MW installation setup
Also changed the name of the log since it's used by all the modules
and not just the explain script

There is a lot of cleanup left to do after this refactor, but
everything appears to work ok, so here's the commit, such as it is.