group hosts with the same table structure on a wiki together
This cuts down a little bit on the output. There's more to be done;
it would be nice, in the case of checking multiple wikis against
one 'master wiki' table structure, if we could report variances
in only one wiki on a host if many wikis it serves have the same
differences. But this can get complicated fast.