Cut down on amountof output when using one dbhost and wiki as main/master
We now group together wikis on a dbhost that have the same structure,
and since we already group together dbhosts that have the same output
for one wiki, we can skip output for wikis and dbhosts with the same
structure as an arbitrary dbhost/wiki in the group.

We also print the full common structure once per group. In some cases
the reason for two separate groups can be as subtle as a NOT NULL
for some column, so this lets the user check the specifics.

Also added params_ignore, a comma-separated list of table parameters
which will be ignored for the purpose of grouping or producing diffs
against the master (both per section master or single master).