use logger in checker script instead of prints for verbosity
This also provides for the table structure for a dbhost/wiki group
to be written only to the log and not to the console, for checks
where a main wiki and dbhost are selected as the basis of comparison.
While having the table structure info is helpful for detailed
manual checks later, it's nicer not to have it mixed into the console
output.