move a couple chunks of code out of the huge display diffs all sects
In the checker, that method was just too long and complicated.
Move the code to display the master table info out to its own method,
move the code to del wth wikis that have the same table descriptions
also out to its own method.
Additionally make DbInfo.get_masters remove dups in the list it returns,
so the caller doesn't have to.